requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
async-timeout==3.0.1
asyncpg==0.18.3
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
discord.py==1.2.2
idna==2.8
multidict==4.5.2
pycparser==2.19
PyNaCl==1.3.0
six==1.12.0
websockets==6.0
yarl==1.3.0
youtube-dl==2019.6.21
